+++ tests/test_query/wikibase/connectivity_obs/__init__.py
