+++ tests/test_query/wikibase/connectivity_obs/connectivity_fragment.py
