+++ tests/test_query/wikibase/connectivity_obs/test_wikibase_connectivity_most_recent_observation_query.py
@@ -2,10 +2,10 @@
 
 from freezegun import freeze_time
 import pytest
-from tests.test_query.test_wikibase_connectivity_observation_query.assert_connectivity_observation import (
+from tests.test_query.wikibase.connectivity_obs.assert_connectivity import (
     assert_connectivity_observation,
 )
-from tests.test_query.test_wikibase_connectivity_observation_query.wikibase_connectivity_observation_fragment import (
+from tests.test_query.wikibase.connectivity_obs.connectivity_fragment import (
     WIKIBASE_CONNECTIVITY_OBSERVATION_FRAGMENT,
 )
 from tests.test_schema import test_schema
