+++ tests/test_query/wikibase/property_popularity_obs/assert_property_popularity.py
