+++ tests/test_query/wikibase/property_popularity_obs/property_popularity_fragment.py
