+++ tests/test_query/wikibase/user_obs/user_fragment.py
